+++ src/components/Apps/LabelsTable.vue
